Added rentRates array to PropertySpace, which
contains the property's cost in its first index, 

and the rentRates according to how many houses/hotels are owned in the
remaining indexes. Made it so this array is passed into PropertySpace
when it is created. Also updated Railroads and Utilities to take in this
array as well, since they inherit from PropertySpace. Also Created
Jagged double array in PropertyGroups class of type PropertySpace to
contain Properties by color to check if all properties of a color are
owned by a single person. Finished checkMonopoly method in PropertyGroup
Class.